ShowPokemon/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.5
Django==4.0.6
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
filelock==3.7.1
idna==3.3
Pillow==9.2.0
platformdirs==2.5.2
pokebase==1.3.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.16.2